AUX change (#90)
error: invalid path 'src/aux.py'
On Windows, certain file names are reserved and cannot be used, even if they appear to be normal filenames. aux is one of these reserved device names, along with others like:

CON

PRN

AUX

NUL

COM1 to COM9

LPT1 to LPT9

So src/aux.py is an invalid path on Windows, which is why Git fails to check out the repository even though it was cloned successfully.